+++ gax/test/unit/streaming.ts
@@ -1157,10 +1157,10 @@ describe('streaming', () => {
       true // new retry behavior enabled
     );
     // resumption strategy is to pass a different arg to the function
-    const getResumptionRequestFn = (originalRequest: RequestType) => {
+    const getResumptionRequestFn = sinon.spy((originalRequest: RequestType) => {
       assert.strictEqual(originalRequest.arg, 0);
       return {arg: 2};
-    };
+    });
     const s = apiCall(
       {arg: 0},
       {
@@ -1194,6 +1194,105 @@ describe('streaming', () => {
         receivedData.join(' '),
         'Hello World testing retries'
       );
+      assert.strictEqual(getResumptionRequestFn.callCount, 1);
+      done();
+    });
+  });
+  it('errors when there is a resumption request function an gaxStreamingRetries is not enabled', done => {
+    // stubbing cancel is needed because PassThrough doesn't have
+    // a cancel method and cancel is called as part of the retry
+    sinon.stub(streaming.StreamProxy.prototype, 'cancel');
+    const error = Object.assign(new GoogleError('test error'), {
+      code: 14,
+      details: 'UNAVAILABLE',
+      metadata: new Metadata(),
+    });
+
+    const spy = sinon.spy((...args: Array<{}>) => {
+      //@ts-ignore
+      const arg = args[0].arg;
+      assert.strictEqual(args.length, 3);
+      const s = new PassThrough({
+        objectMode: true,
+      });
+      switch (arg) {
+        case 0:
+          s.push('Hello');
+          s.push('World');
+          setImmediate(() => {
+            s.emit('metadata');
+          });
+          setImmediate(() => {
+            s.emit('error', error);
+          });
+          setImmediate(() => {
+            s.emit('status');
+          });
+          return s;
+        case 1:
+          s.push(null);
+          setImmediate(() => {
+            s.emit('error', new Error('Should not reach'));
+          });
+
+          setImmediate(() => {
+            s.emit('status');
+          });
+          return s;
+        case 2:
+          s.push('testing');
+          s.push('retries');
+          setImmediate(() => {
+            s.emit('metadata');
+          });
+          setImmediate(() => {
+            s.emit('end');
+          });
+          return s;
+        default:
+          setImmediate(() => {
+            s.emit('end');
+          });
+          return s;
+      }
+    });
+    const apiCall = createApiCallStreaming(
+      spy,
+      streaming.StreamType.SERVER_STREAMING,
+      false,
+      false // new retry behavior disabled
+    );
+    // resumption strategy is to pass a different arg to the function
+    const getResumptionRequestFn = (originalRequest: RequestType) => {
+      assert.strictEqual(originalRequest.arg, 0);
+      return {arg: 2};
+    };
+    const s = apiCall(
+      {arg: 0},
+      {
+        retry: gax.createRetryOptions(
+          [14],
+          {
+            initialRetryDelayMillis: 100,
+            retryDelayMultiplier: 1.2,
+            maxRetryDelayMillis: 1000,
+            rpcTimeoutMultiplier: 1.5,
+            maxRpcTimeoutMillis: 3000,
+            maxRetries: 2, // max retries or timeout must be > 0 in order to reach the code we want to test
+          },
+          undefined,
+          getResumptionRequestFn
+        ),
+      }
+    );
+
+    s.on('error', err => {
+      // double check it's the expected error on the stream
+      // stream will continue after retry
+      assert.deepStrictEqual(
+        err.message,
+        'getResumptionRequestFn can only be used when gaxStreamingRetries is set to true.'
+      );
       done();
     });
   });
@@ -1205,6 +1304,67 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
   });
 
   it('server streaming call retries until exceeding max retries', done => {
+    const retrySpy = sinon.spy(streaming.StreamProxy.prototype, 'retry');
+    const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
+      code: 14,
+      details: 'UNAVAILABLE',
+      metadata: new Metadata(),
+    });
+
+    const spy = sinon.spy((...args: Array<{}>) => {
+      assert.strictEqual(args.length, 3);
+      const s = new PassThrough({
+        objectMode: true,
+      });
+      setImmediate(() => {
+        s.emit('metadata');
+      });
+      setImmediate(() => {
+        s.emit('error', firstError);
+      });
+      return s;
+    });
+
+    const apiCall = createApiCallStreaming(
+      spy,
+      streaming.StreamType.SERVER_STREAMING,
+      false,
+      true
+    );
+
+    const call = apiCall(
+      {},
+      {
+        retry: gax.createRetryOptions([14], {
+          initialRetryDelayMillis: 100,
+          retryDelayMultiplier: 1.2,
+          maxRetryDelayMillis: 1000,
+          rpcTimeoutMultiplier: 1.5,
+          maxRpcTimeoutMillis: 3000,
+          maxRetries: 2,
+        }),
+      }
+    );
+
+    call.on('error', err => {
+      try {
+        assert(err instanceof GoogleError);
+        if (err.code !== 14) {
+          // ignore the error we are expecting
+          assert.strictEqual(err.code, 4);
+          assert.strictEqual(retrySpy.callCount, 2);
+          assert.strictEqual(
+            err.message,
+            'Exceeded maximum number of retries before any response was received'
+          );
+          done();
+        }
+      } catch (error: unknown) {
+        done(error);
+      }
+    });
+  });
+  it('does not retry when there is no shouldRetryFn and retryCodes is an empty array', done => {
     const retrySpy = sinon.spy(streaming.StreamProxy.prototype, 'retry');
     const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
       code: 14,
@@ -1237,7 +1397,8 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
     const call = apiCall(
       {},
       {
-        retry: gax.createRetryOptions([14], {
+        // pass an empty array for retryCodes
+        retry: gax.createRetryOptions([], {
           initialRetryDelayMillis: 100,
           retryDelayMultiplier: 1.2,
           maxRetryDelayMillis: 1000,
@@ -1250,17 +1411,9 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
 
     call.on('error', err => {
       assert(err instanceof GoogleError);
-      if (err.code !== 14) {
-        // ignore the error we are expecting
-        assert.strictEqual(err.code, 4);
-        // even though max retries is 2
-        // the retry function will always be called maxRetries+1
-        // the final call is where the failure happens
-        assert.strictEqual(retrySpy.callCount, 3);
-        assert.strictEqual(
-          err.message,
-          'Exceeded maximum number of retries before any response was received'
-        );
+      if (err.code === 14) {
+        assert.strictEqual(err.code, 14);
+        assert.strictEqual(retrySpy.callCount, 0);
         done();
       }
     });